refactor(custom-widgets): satisfy architecture checks
